\name{c}
\alias{c}
\alias{c.disord}
\alias{c,disord-method}
\title{Concatenation}
\description{
  Concatenation simply does not make sense for \code{disord} objects.}
\value{Returns an error.}
\author{Robin K. S. Hankin}
\note{
  I could not figure out how to stop idiom like \dQuote{\code{c(1,rdis())}} from
  returning a result.  Just don't use it, OK?
}
