\name{Extract}
\alias{[.disord}
\alias{[<-.disord}
\alias{[,disord-method}
\alias{[,disord,index,ANY-method}
\alias{[,disord,index,missing-method}
\alias{[<-,disord-method}
\alias{[,disord,ANY,ANY-method}
\alias{[<-,disord,ANY,ANY-method}
\alias{[<-,disord,index,ANY,ANY-method}
\alias{[<-,disord,index,missing,ANY-method}
\alias{[<-,disord,index,missing,numeric-method}
\alias{[<-,disord,index,missing,disord-method}
\alias{[<-,disord,missing,missing,numeric-method}
\alias{[<-,disord,missing,missing,disord-method}
\alias{[,disord,index,ANY,ANY-method}
\alias{[,disord,index,missing,ANY-method}
\alias{[}
\alias{[<-}
\title{Extract or Replace Parts of disords}
\description{
   Methods for \code{"["} and  \code{"[<-"}, i.e., extraction or
   subsetting of disord objects.
 }
 \usage{
x[i]
x[i] <- value
}
\arguments{
  \item{x}{Object of class disord}
  \item{index}{replacement value}
  \item{value}{replacement value}
}
\value{
  Extraction and methods return a disord.
}
\author{Robin K. S. Hankin}
}
\examples{
2+2
}
